### 2024-12-02 #7548 @editorialbot
[PRE REVIEW]: eva3dm: A R-package for model evaluation of 3D weather and air quality models
**Submitting author:** @Schuch666 (<a href="http://orcid.org/0000-0001-5977-4519">Daniel Schuch</a>)
**Repository:** https://github.com/Schuch666/eva3dm
**Branch with paper.md** (empty if default branch): JOSS
**Version:** 1.00
