### 2024-12-02 #7548 @kthyng
@schuch666:
* I see the MIT license badge but this is the license file I have found: https://github.com/Schuch666/eva3dm/blob/main/LICENSE — do you have the MIT license file somewhere in your repo?
* Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly. (yours is presently 1268 words)
* Do you list installation instructions somewhere?
