### 2024-12-02 #7548 @Schuch666
Hi @kthyng , no. I have tried other formats before submit to CRAN, but this is the format that CRAN require to publish in the official R repository. Do I need to create a different license file or format?

I will reduce the word count (I did not find this information before submit).

The installation instructions are in the read-me from the JOSS branche (I will update the main branch in the future) and it should be just one command to install from CRAN or github, the package has only two main dependencies that are installed from CRAN automatically. Let me know if I need to include more documentation on the installation instructions.
