+++ paper.md
@@ -19,33 +19,33 @@ bibliography: paper.bib
 
 # Summary
 
-Eva3dm is a package designed to support the evaluation of 3-dimensional physical models (in special, weather and air quality models) against observation data in order to quantify different errors and bias present in the model results and/or classify the model performance based on criteria from the literature.
+Eva3dm is a package designed to support the evaluation of 3-dimensional physical models (in special, weather and air quality models) against observation data in order to quantify different errors and bias present in the model results.
 
 # Statement of need
 
-Evaluation is the most important step for any model application, it can not only assure that the model results represent (or not) accurately the interest variable but also can guide to choose between different model configurations and guide model development. There are currently other tools available in R [@David:2012], Python [@Ladwig:2017] or other languages [@NCAR:2019;@Appel:2011] but none provide the degree of integrated support presented in this package: from the pre-processing of observations and model output, evaluation and visualization.
+Evaluation is the most important step for any model application, it can assure that the model results represent accurately the interest variables. There are currently other tools available in R [@David:2012], Python [@Ladwig:2017] or other languages [@NCAR:2019;@Appel:2011] but none provide the degree of integrated support presented in this package: from the pre-processing of observations and model output, evaluation and visualization.
 
 # Description
 
-In the literature there are many different evaluation criteria depending on the application and variable [@Emery:2001;@Ramboll:2018;@Monk:2019;@Zhang:2019;@Emery:2017;@Zhai:2024]. These criteria varies depending on the goal of the simulation, the variability of the observations, the error associated to the measurements, and how other models available in the literature compared with observations and should be used accordingly to the particular application, study region and other uncertainties present on the model and model inputs.
+In the literature there are many different evaluation criteria depending on the application and variable [@Emery:2001;@Ramboll:2018;@Monk:2019;@Zhang:2019;@Emery:2017;@Zhai:2024]. These criteria varies depending on the goal of the simulation, the variability of the observations, the error associated to the measurements, and how other models available in the literature compared with observations and should be used accordingly to the particular application.
 
-The package was designed to support the pre-processing of observations, post-processing of model output, evaluation and visualization of the results from the evaluation. A brief description of these 4 steps is provided bellow:
+A brief description of the steps to perform a model evaluation is provided bellow:
 
 **1. Pre-processing of observations**: 
 
 - Download of the observations, example of observations available: 
-  - METAR of meteorological observations can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or via the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) site
-  - AERONET data for Aerosol Optical Depth can be downloaded at [AErosol RObotic NETwork](https://aeronet.gsfc.nasa.gov/new_web/data.html)
+  - METAR observations can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) site
+  - AERONET can be downloaded at [AErosol RObotic NETwork](https://aeronet.gsfc.nasa.gov/new_web/data.html) site
   - Air Quality data for Brazil can be downloaded using the R-package [qualR](https://github.com/ropensci/qualR), or [QUALAR](https://qualar.cetesb.sp.gov.br/qualar) and [MonitorAir](https://www.data.rio/datasets/dados-hor%C3%A1rios-do-monitoramento-da-qualidade-do-ar-monitorar/explore) sites
   - Satellite products are available at [NASA giovanni](https://giovanni.gsfc.nasa.gov/giovanni/) website
 
-- Process a list of the location to extract time-series from the model. NOTE: There are examples of list of sites in the documentation of `extract_serie` for METAR, AERONET, INMET in Brazil, and Air Quality stations in Brazil.
[…]
