### 2024-12-02 ef03d3a
reducing words
