### 2024-12-02 a4ad997
reducing words
