+++ paper.md
@@ -33,17 +33,17 @@ A brief description of the steps to perform a model evaluation is provided bello
 
 **1. Pre-processing of observations**: 
 
-- Download of the observations, example of observations available: 
-  - METAR observations can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) site
+- Download of the observations, some examples include: 
+  - METAR can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) site
   - AERONET can be downloaded at [AErosol RObotic NETwork](https://aeronet.gsfc.nasa.gov/new_web/data.html) site
   - Air Quality data for Brazil can be downloaded using the R-package [qualR](https://github.com/ropensci/qualR), or [QUALAR](https://qualar.cetesb.sp.gov.br/qualar) and [MonitorAir](https://www.data.rio/datasets/dados-hor%C3%A1rios-do-monitoramento-da-qualidade-do-ar-monitorar/explore) sites
   - Satellite products are available at [NASA giovanni](https://giovanni.gsfc.nasa.gov/giovanni/) website
 
 - Process a list of the location to extract time-series from the model.
 
-- Process observation data for evaluation: Calculation of secondary variables.
+- Process observation data for evaluation: Unit conversion, time zone conversion to UTC, and Calculation of secondary variables.
 
-- Quality Assurance of the observation data: in this point the user should check for values that are outside the range of each variable, check if there is actual data from the time-period and region of the simulation and note any singular event that can be represented.
+- Quality Assurance of the observation data: check for values outside the valid range, check if the data is available for the time-period and region of the simulation and note any singular event.
 
 **2. Pre-processing of model output**: Extraction and post-processing of model outputs using `evad3m` functions and/or other tools and calculation of secondary variables.
 
@@ -51,23 +51,23 @@ A brief description of the steps to perform a model evaluation is provided bello
 
 **4. Visualization**: There are functions to visualize model fields (taken into account different projections) and overlay of the statistical results.
 
-To support the steps 1 to 4, the R-package includes the following functions:
+To support these steps, following functions are available:
 
 ### Pre-processing of observations:
 
-The functions `rh2q` and `q2rh` converts humidity units and the functions `mda8`, `ma8h`, `hourly`, and `daily` can be used to calculate time-series from observations and model. Most of the processing should be done by the user since the formats and conventions are different depending on the source of the data. The format used to evaluate time-series is a data.frame, the first column must contain time (in POSIXlt) and one additional column for each different location.
+The functions `rh2q` and `q2rh` converts humidity units and the functions `mda8`, `ma8h`, `hourly`, and `daily` can be used to calculate average of time-series. The format used to evaluate time-series is a data.frame, the first column must contain time (in POSIXlt) and one additional column for each different location, satellite data can be read using the function `rast` form terra package.
 
 ### Pre-processing of model outputs:
 
-The function `extract_serie` extract and save time-series from model outputs (or other compatible NetCDF files) using a data.frame with name (row names) and latitude (column lat) and longitude (column lon), while the functions `extract_mean` and `extract_max_8h` extract the average or the daily maximum of 8-hour moving average and save in a new NetCDF file;
+The function `extract_serie` extract and save time-series from model outputs using a data.frame with name (row names) and latitude (column lat) and longitude (column lon), while the functions `extract_mean` and `extract_max_8h` extract the average or the daily maximum of 8-hour moving average and save in a new NetCDF file;
 
[…]
