### 2024-12-02 #7548 @Schuch666
@kthyng the number of word is fixed
