### 2024-12-10 #7548 @kthyng
Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "r" and "python", but please check for others.
