### 2024-12-12 #7548 @kthyng
@Schuch666 We are not requiring a second LICENSE file for now (though may in the future, we need to discuss). Thanks.
