### 2024-12-12 #7548 @Schuch666
Hi, @kthyng  no problem, thank you

NOTE: the users tomgrylls, mvreeuwijk, madDipanjan and other developers from uDALES model can be potential additional reviewers (if they are available), the publication is related to their paper [uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling](https://doi.org/10.21105/joss.03055)
