### 2024-12-19 #7548 @Schuch666
Hi @kthyng , what is the status of the review process? do you think it will resume in a few day or after the holidays?
