### 2025-02-11 d0a3fc6
JOSS update, editor sugestions
