### 2025-02-11 53f3943
Figure caption update
