### 2025-02-11 #7548 @Schuch666
Hi @elbeejay
Thank you for the suggestions
This new version of the paper is fixed, including all the suggestions
Please, l;et me know if you have any additional questions or suggestions
