### 2025-02-15 #7548 @elbeejay
@adithirugis will you consider reviewing this submission to JOSS titled "eva3dm: A R-package for model evaluation of 3D weather and air quality models"? Given your experience with air quality data, as well as your previous experience publishing an R Shiny package "mmaqshiny" with JOSS, I thought you might be a good fit.

I'm guessing you are familiar with JOSS style reviews, but I will just ask that you review our [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), prior to accepting this invitation to review. We are currently asking reviewers to complete their reviews within 6 weeks of the review issue starting, but can be flexible as needed. Please let me know if you'd be interested in reviewing, thanks!

Jay
