### 2025-02-16 #7548 @elbeejay
@zhenchun will you consider reviewing this submission to JOSS titled "eva3dm: A R-package for model evaluation of 3D weather and air quality models"? I know you've reviewed for JOSS in the past, so I won't belabor our practices. I did think you might be a good fit for this submission due to your experience with both R and your work in air pollution and air quality.

Please take a look and let me know if you'd be interested. We are currently asking reviewers to complete their reviews within 6 weeks of the review issue starting, but can be flexible as needed. No pressure if you're caught up with other things, we appreciate the volunteer work you've done for JOSS in the past!

Jay
