### 2025-02-17 #7548 @elbeejay
@jonsampedro would you consider reviewing this submission, titled "eva3dm: A R-package for model evaluation of 3D weather and air quality models", for JOSS? I know you've published in JOSS on R packages that fetch and manipulate air pollution data, so I thought you might be a good fit as a peer-reviewer for this one.

Please take a look at the repository and let us know if you think you'll be able to peer-review. Right now we're asking reviewers to try to complete their reviews within 6 weeks, however we can be flexible as needed.

Thanks,
Jay
