### 2025-02-17 #7548 @elbeejay
@cvitolo, would you consider reviewing this submission, titled "eva3dm: A R-package for model evaluation of 3D weather and air quality models", for JOSS? I see that you've published multiple R packages in JOSS in the past, including your `rdefra` library which improves access to UK air pollution data. That being the case, I thought you might be able to provide valuable suggestions and comments to our submitting authors here.

Please take a look at the repository and let us know if you think you'll be able to peer-review. Right now we're asking reviewers to try to complete their reviews within 6 weeks, however we can be flexible as needed.

Thanks,
Jay
