### 2025-02-19 #7797 @elbeejay
@zhenchun, @jonsampedro thank you again for agreeing to peer-review for JOSS. This is the main review issue associated with this submission. Instructions are provided at the top of this issue on how to generate your reviewer checklist, and how to conduct your reviews. Should you have any questions, please do not hesitate to reach out to me.

It'd be great if you could aim to complete your first reviews in 6 weeks time (by April 2) - I will have the bot send out reminders in 3 weeks to try and help keep us all on track. If that timeline is not going to work for you, please let us know here so we can figure out a plan that is feasible for everyone.
