### 2025-02-20 8e7abd3
suppress the original axis if lat and long are set to false
