### 2025-02-20 1e14f34
added feature to extract time series from diff model levels
