+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-02-20T21:40Z
+last_built: 2025-02-20T21:53Z
 urls:
   reference: https://schuch666.github.io/eva3dm/reference
   article: https://schuch666.github.io/eva3dm/articles
