### 2025-03-10 #7797 @jonsampedro
Hello @elbeejay, @Schuch666,

I have reviewed the package and I think it is a valuable tool for comparing climate and pollution observations with model results. It includes various functions for evaluating model performance, along with visualizations of key findings. While I haven't encountered any significant issues with the package's functionality, I believe it would benefit from a more comprehensive documentation website and paper. More specific comments are provided below.

### **Functionality**
I have tested the main evaluation functions (`eva` and `sat`) using the sample data provided with the package and have not encountered any issues. The output visualization functions (e.g., `plot_overlay`) have also worked correctly.
However, I did encounter some errors when testing and checking the package, which I believe are related to dependencies, specifically with the `add_box` and `add_legend` functions from the `terra` package.

### **Documentation and manuscript suggestions**
The package includes a documentation webpage with a basic workflow description and a list of functions. Enhancing the documentation site would improve understanding of the package’s full functionality and encourage wider adoption. Here are some suggestions:
- Add an introduction/summary and clear installation guidelines.
- Expand on the different stages of the workflow, perhaps by creating individual pages or vignettes for each stage.
- Describe the data provided with the package in the "Pre-processing of observations".
- While the examples included within each function are helpful, a complete “step-by-step” example on the documentation site would significantly improve understanding of the package’s capabilities.

I also believe the **manuscript** could be improved in several ways:
- The "Statement of Need" is quite brief and could be expanded to better reflect the importance of evaluating climate and air pollution models in current research.
- To improve the flow of the Description section, the explanation of the four steps in the evaluation process could include the package functions discussed later in the text.
- Including a schematic figure outlining the workflow and steps would help clarify the overall process.
- Adding some example code would make the manuscript more practical and easier to follow.

### **Additional Comments**
- There is a reference (badge) to an MIT-type license, but the file does not follow the standard structure outlined here: https://opensource.org/license/mit.
- The package does not include a Contributing file with community guidelines. Adding one would help encourage contributions and clarify the process for potential contributors.
- The authors mention that the package could be applied to different models beyond WRF-Chem. Providing additional guidance on how to adapt the package to other models would be highly valuable for the community.
