+++ README.md
@@ -32,6 +32,7 @@ the last version can be installed directly from github (using the `remotes` pack
 
 All the functions include examples in the [documentation](schuch666.github.io/eva3dm).
 
+## Recomended Workflow
 The following workflow is recommended:
 
 **1. Pre-processing of observations**: 
