+++ vignettes/workflow.Rmd
@@ -20,18 +20,20 @@ library(riem)
 library(terra)
 ```
 
-In this Vignette a example of evaluation of temperature at surface level using data from METeorological Aerodrome Report (METAR).
+The recommended workflow to evaluate your model results will include 4 steps:
 
-The recommended workflow to evaluate your model results will include 4 steps
+
 1. Pre-processing of observations
+
 2. Pre-processing of model output
+
 3. Model Evaluation
+
 4. Visualization
 
-The exact steps needed to evaluation will will be different for different models, different observation data-set, and different variable. However, in this example a complete workflow to evaluate the model using surface stations for one variable.
+The exact steps needed to evaluation will be different for different models, different observation data-sets, and different variables. However, in this vignette is presented an example of evaluation of temperature at surface level (T2) from WRF-Chem for a simulation at 3-km of resolution for January 2016 using using data from METeorological Aerodrome Report (METAR).
 
-1. Pre-processing of observations:
-
+1 - Pre-processing of observations:
 We are using the r-package `riem` to download METAR data:
 
 ```{r metar}
@@ -61,40 +63,41 @@ for(site in sites){
 }
 
 ```
+Lets check if the observations are ok
 
-QA of the observation data.
-
-```{r check}
+```{r check, fig.width = 7, fig.height = 4}
 plot(METAR$date, METAR[,2], ty = 'l',xlab = '',ylab = 'T2', main = 'METAR OBS')
 head(METAR)
 ```
-In this case the temperature units need to be converted from Fahrenheit to Celsius and the data need to be grouped or averaged by exact hours, in this case using the hourly function.
-```{r observation}
+In this case the temperature units needs to be converted from Fahrenheit to Celsius and the data needs to be grouped or averaged by exact hours, in this case using the hourly function.
+```{r observation, fig.width = 7, fig.height = 4}
 
 METAR[,-1] <- 5/9 * (METAR[,-1]-32)
 METAR      <- hourly(METAR)
 
-plot(METAR$date, METAR[,2], ty = 'l',xlab = '',ylab = 'T2', main = 'METAR OBS')
+plot(METAR$date, METAR[,2], ty = 'l',xlab = '',ylab = 'T2', main = 'METAR processed OBS')
 head(METAR)
 
 ```
 
-For the next step a site-list with latitude and logitude of each surfacve station is needed, a list of all METAR stations is avaliable in eva3dm package.
+For the next step a site-list with latitude and longitude of each surface station is needed, a list of all METAR stations is available in eva3dm package.
 ```{r site-list}
 site_list <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/sites_METAR.Rds"))
 head(site_list)
 ```
 
-2. Pre-processing of model output: model output are large files and and do not allow the direct comparison with observations. The package provide a function that extract time-series from geogrephic locations from a site-list. The function `extract_serie()` extract the time-series from a list of files and save the output in R file (in this example the output is read). 
+2 - Pre-processing of model output:
+The model output are large files and do not allow the direct comparison with observations. The package provides a function that extracts time-series from geographic locations from a site-list. The function `extract_serie()` extracts the time-series from a list of files and saves the output in R file (in this example we are reading the R file generated by `extract_serie()`).
 ```{r model}
 ## to extract time-series from WRF-Chem model
 ## wrf_files <- dir(pattern = "wrfout_d01")
 ## extract_serie(filelist = wrf_files, point = site_list, variable="T2", prefix="mmodel.d01", field="3d")
[…]
