+++ vignettes/workflow.Rmd
@@ -87,12 +87,13 @@ head(site_list)
 ```
 
 2 - Pre-processing of model output:
-The model output are large files and do not allow the direct comparison with observations. The package provides a function that extracts time-series from geographic locations from a site-list. The function `extract_serie()` extracts the time-series from a list of files and saves the output in R file (in this example we are reading the R file generated by `extract_serie()`).
+The model output are large files and do not allow the direct comparison with observations. The package provides a function that extracts time-series from geographic locations from a site-list. The function `extract_serie()` extracts the time-series from a list of files and saves the output in R file (in this example we are reading the R file generated by `extract_serie()`). An finally, the temperature extracted from the model is converted from Kelvin to Celsius.
 ```{r model}
 ## to extract time-series from WRF-Chem model
 ## wrf_files <- dir(pattern = "wrfout_d01")
 ## extract_serie(filelist = wrf_files, point = site_list, variable="T2", prefix="mmodel.d01", field="3d")
 model_d01 <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/model.d01.T2.Rds"))
+model_d01[-1] <- model_d01[-1] - 273.15
 ```
 The function `wrf_rast()` should be used, instead of `extract_serie()`, to extract the model output to evaluate against observations in regular grids (i.e. satellite products).
 
@@ -119,4 +120,4 @@ terra::lines(BR)
 terra::lines(masp, col = 'gray')
 legend_range(spatial_table$MB,y = table["ALL","MB"])
 ```
-The WRF-Chem temperature in this example showed a very low MB, except for the SBSJ station (MB=-0.7), with are in the criteria for acceptable performance (|MB| < 0.5) and the criteria for evaluation for complex terrain for the SBSJ case (|MB| < 1.0). Another criteria for temperature are IAO > 0.8 and ME < 2.0 (ME < 3, for complex terrain), in general the model presented a reasonable performance for temperature. More information and references can be found on the documentation of `stat()` function.
+The WRF-Chem temperature in this example showed a very low MB, except for the SBSJ station (MB=-0.7), which is in the criteria for acceptable performance (|MB| < 0.5) and the criteria for evaluation for complex terrain for the SBSJ case (|MB| < 1.0). Another criteria for temperature are IAO > 0.8 and ME < 2.0 (ME < 3, for complex terrain), in general the model presented a reasonable performance for temperature. More information and references can be found on the documentation of `stat()` function.
