+++ vignettes/workflow.Rmd
@@ -95,7 +95,7 @@ The model output are large files and do not allow the direct comparison with obs
 model_d01 <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/model.d01.T2.Rds"))
 model_d01[-1] <- model_d01[-1] - 273.15
 ```
-The function `wrf_rast()` should be used, instead of `extract_serie()`, to extract the model output to evaluate against observations in regular grids (i.e. satellite products).
+NOTE: to extract time-series from other models (not WRF or WRF-Chem), the arguments `latitude` and `longitude` from the function `extract_serie()` need to match the name of these variables for latitude and longitude in the model output. If the model do not include the Times variable, `use_TFLAG` must be set to `TRUE` to extract and convert the time from TFLAG variable (for example for CMAQ / CAMx models) or `use_datesec` must be set to `TRUE` to extract and convert time from the datasec variable (for example for WACCM / CAM-Chem models).
 
 3 - Model Evaluation: 
 The functions `eva()` can be used to calculate the statistical indexes from time-series for individual stations or all stations combined.
@@ -107,10 +107,9 @@ for(site in sites){
 table <- eva(mo = model_d01, ob = METAR, site = 'ALL', table = table)
 print(table)
 ```
-To evaluate using observations in regular grids, the function `sat()` is used instead of `eva()`.
 
 4 - Visualization:
-The function `overlay()` can be used to plot the results from `eva()` or `sat()` on a map. First the data is georeferenced with `%at%`.
+The function `overlay()` can be used to plot the results from `eva()` on a map. First the data is georeferenced with `%at%`.
 ```{r visualize, fig.width = 7, fig.height = 5.5}
 spatial_table <- table %at% site_list
 overlay(spatial_table, z = 'MB', main = 'T2 main bias (MB)',expand = 1.6,lim = 0.1)
