### 2025-03-24 01776c9
updates on vignete and paper
