+++ paper.md
@@ -24,19 +24,19 @@ Eva3dm is a package designed to support the evaluation of 3-dimensional physical
 
 # Statement of need
 
-Evaluation is the most important step for any model application, it can assure that the model results represent accurately the interest variables. There are currently other tools available in R [@David:2012], Python [@Ladwig:2017] or other languages [@NCAR:2019;@Appel:2011] but none provide the degree of integrated support presented in this package: from the pre-processing of observations and model output, to evaluation and visualization.
+Evaluation is a crucial step in any model application, as it ensures that the model results accurately represent the variables of interest. Without a good evaluation process, the reliability and applicability of model outputs remain uncertain. There are currently other tools available in R [@David:2012], Python [@Ladwig:2017] or other languages [@NCAR:2019;@Appel:2011]. However, these tools often focus on specific aspects and lack a fully integrated framework. This package fills that gap by streamlining the entire evaluation process—from preprocessing observations and model outputs to statistical analysis and visualization—offering a comprehensive and user-friendly solution for air quality model assessment.
 
 # Description
 
 The literature presents various evaluation criteria depending on the variable [@Emery:2001;@Ramboll:2018;@Monk:2019;@Zhang:2019;@Emery:2017;@Zhai:2024], which can be used to compare models and assess their performance. These criteria vary based on the simulation goal, observation variability, and measurement errors.
 
-A brief description of the steps to perform a model evaluation is provided below:
+A brief description of the steps to perform a model evaluation and the functions to support these steps are descripted on the next sessions.
 
-**1. Pre-processing of observations**: 
+### 1. Pre-processing of observations
 
 - Download of the observations, some examples include: 
-  - METAR can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) site
-  - AERONET can be downloaded at [AErosol RObotic NETwork](https://aeronet.gsfc.nasa.gov/new_web/data.html) site
+  - METAR (METeorological Aerodrome Report) can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) site
+  - AERONET (Aerosol Robotic Network) can be downloaded at [AErosol RObotic NETwork](https://aeronet.gsfc.nasa.gov/new_web/data.html) site
   - Air Quality data for Brazil can be downloaded using the R-package [qualR](https://github.com/ropensci/qualR), or [QUALAR](https://qualar.cetesb.sp.gov.br/qualar) and [MonitorAir](https://www.data.rio/datasets/dados-hor%C3%A1rios-do-monitoramento-da-qualidade-do-ar-monitorar/explore) sites
   - Satellite products are available at [NASA giovanni](https://giovanni.gsfc.nasa.gov/giovanni/) website
 
@@ -46,19 +46,11 @@ A brief description of the steps to perform a model evaluation is provided below
 
 - Quality Assurance of the observation data: check for values outside the valid range, check if the data is available for the time-period and region of the simulation and note any singular event.
 
-**2. Pre-processing of model output**: Extraction of model outputs, unit conversion and calculation of secondary variables.
-
-**3. Model Evaluation**: The evaluation involves pairing observations with model results and calculating the statistical and/or categorical indices.
-
-**4. Visualization**: Visualization of model results and statistical results.
-
-To support these steps, following functions are available:
-
-### Pre-processing of observations:
-
[…]
