### 2025-03-25 #7797 @elbeejay
Checking in - @zhenchun have you had a chance to test the functionality of the package and otherwise read and review the documentation and paper?

@Schuch666 how are the revisions going in response to @jonsampedro's comments? I see some changes have been made in the JOSS branch to the paper itself.
