+++ vignettes/workflow.Rmd
@@ -31,7 +31,7 @@ The recommended workflow to evaluate your model results will include 4 steps:
 
 4. Visualization
 
-The exact steps needed to evaluation will be different for different models, different observation data-sets, and different variables. However, in this vignette is presented an example of evaluation of temperature at surface level (T2) from WRF-Chem for a simulation at 3-km of resolution for January 2016 using using data from METeorological Aerodrome Report (METAR).
+The exact steps needed for evaluation will be different for different models, different observation data-sets, and different variables. However, in this vignette is presented an example of evaluation of temperature at surface level (T2) from WRF-Chem for a simulation at 3-km of resolution for January 2016 using data from METeorological Aerodrome Report (METAR).
 
 1 - Pre-processing of observations:
 We are using the r-package `riem` to download METAR data:
@@ -69,7 +69,7 @@ Lets check if the observations are ok
 plot(METAR$date, METAR[,2], ty = 'l',xlab = '',ylab = 'T2', main = 'METAR OBS')
 head(METAR)
 ```
-In this case the temperature units needs to be converted from Fahrenheit to Celsius and the data needs to be grouped or averaged by exact hours, in this case using the hourly function.
+In this case the temperature units needs to be converted from Fahrenheit to Celsius and the data needs to be grouped or averaged in time, in this case we are using the `hourly()` to perform a evaluation against hourly data.
 ```{r observation, fig.width = 7, fig.height = 4}
 
 METAR[,-1] <- 5/9 * (METAR[,-1]-32)
@@ -87,13 +87,13 @@ head(site_list)
 ```
 
 2 - Pre-processing of model output:
-The model output are large files and do not allow the direct comparison with observations. The package provides a function that extracts time-series from geographic locations from a site-list. The function `extract_serie()` extracts the time-series from a list of files and saves the output in R file (in this example we are reading the R file generated by `extract_serie()`). An finally, the temperature extracted from the model is converted from Kelvin to Celsius.
+The model output are large files and does not allow the direct comparison with observations. The package provides a function that extracts time-series from geographic locations from a site-list. The function `extract_serie()` extracts the time-series from a list of files and saves the output in R file (in this example we are reading the R file generated by `extract_serie()`). An finally, the temperature extracted from the model is converted from Kelvin to Celsius.
 ```{r model}
 ## to extract time-series from WRF-Chem model
-## wrf_files <- dir(pattern = "wrfout_d01")
-## extract_serie(filelist = wrf_files, point = site_list, variable="T2", prefix="mmodel.d01", field="3d")
-model_d01 <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/model.d03.T2.Rds"))
-model_d01[-1] <- model_d01[-1] - 273.15
+## wrf_files <- dir(pattern = "wrfout_d03")
+## extract_serie(filelist = wrf_files, point = site_list, variable="T2", prefix="model.d03", field="3d")
+model_d03 <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/model.d03.T2.Rds"))
+model_d03[-1] <- model_d03[-1] - 273.15
 ```
[…]
