### 2025-03-26 #7797 @Schuch666
Hello @jonsampedro,

All the suggestions was used to improve the package documentation and paper, in special a new vignette was created to shows the steps of the evaluation including a note about evaluation of other models (CMAQ, CAMx, WACCM and CAM-Chem).

See bellow my answers for the questions

Functionality: add_box and add_legend are only exported from terra package, but I can contact the developers if you provide the steps to reproduce the errors.

Website / documentation:
- Introduction/summary and installation guidelines: all added to the package website / read-me. The package should install from CRAN or github with one command.
- A new vignette showing the evaluation for temperature from WRF-Chem using METAR was added to the package.
- The package don't provide observations, but there are some suggestions of datasets that can be used, there are too many data-sets available to cover all, but some additional information was included.
- A new vignette include a step-by-step workflow now, it will be included in the website.

Manuscript:
- "Statement of Need" was updated
- updated Description session to improve the paper flow
- A new figure was included with a schematic of model evaluation workflow
- An simple example is included before the special functions

Additional comments:
- added a new MIT license file, the previous LICENSE file is the CRAN requirements for R-packages
- added a CONTRIBUTION file
- added a NOTE on the vignette
