### 2025-03-26 #7797 @Schuch666
Hello @elbeejay and @jonsampedro
The JOSS brand contains the updated paper, package and documentation
Let me know if you have any question or if the number of words need to be reduced
