### 2025-03-26 437409c
Delete doc directory
