+++ doc/workflow.Rmd
@@ -1,116 +1,0 @@
----
-title: "Introduction to eva3dm"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Introduction to eva3dm}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-```{r setup}
-library(eva3dm)
-library(riem)
-library(terra)
-```
-
-In this Vignette a example of evaluation of temperature at surface level using data from METeorological Aerodrome Report (METAR).
-
-The recommended workflow to evaluate your model results will include 4 steps
-1. Pre-processing of observations
-2. Pre-processing of model output
-3. Model Evaluation
-4. Visualization
-
-The exact steps needed to evaluation will will be different for different models, different observation data-set, and different variable. However, in this example a complete workflow to evaluate the model using surface stations for one variable.
-
-1. Pre-processing of observations:
-
-We are using the r-package `riem` to download METAR data:
-
-```{r metar}
-start_date  <- "2016-01-01"
-end_date    <- "2016-02-01"
-sites       <- c("SBGR","SBKP","SBMT","SBSJ","SBSP","SBST","SBTA")
-
-METAR  <- data.frame(date = seq.POSIXt(as.POSIXct(start_date), 
-                                       as.POSIXct(end_date), 
-                                       by = "hour"))
-for(site in sites){
-  cat("downloading METAR from:",site,"...\n")
-
-  DATA <- riem_measures(station    = site,
-                        date_start = start_date,
-                        date_end   = end_date,
-                        elev       = FALSE,
-                        latlon     = FALSE)
-  DATA <- data.frame(date = DATA$valid,
-                     T2   = DATA$tmpf)
-  names(DATA) <- c('date', site)
-  METAR       <- merge(x     = METAR, 
-                       y     = DATA, 
-                       by    = "date", 
-                       all   = T, 
-                       sort  = TRUE)
-}
-
-```
-
-QA of the observation data.
-
-```{r check}
-plot(METAR$date, METAR[,2], ty = 'l',xlab = '',ylab = 'T2', main = 'METAR OBS')
-head(METAR)
-```
-In this case the temperature units need to be converted from Fahrenheit to Celsius and the data need to be grouped or averaged by exact hours, in this case using the hourly function.
-```{r observation}
-
-METAR[,-1] <- 5/9 * (METAR[,-1]-32)
-METAR      <- hourly(METAR)
-
-plot(METAR$date, METAR[,2], ty = 'l',xlab = '',ylab = 'T2', main = 'METAR OBS')
-head(METAR)
-
-```
-
-For the next step a site-list with latitude and logitude of each surfacve station is needed, a list of all METAR stations is avaliable in eva3dm package.
-```{r site-list}
-site_list <- readRDS(paste0(system.file("extdata",package="eva3dm"),"/sites_METAR.Rds"))
-head(site_list)
-```
-
-2. Pre-processing of model output: model output are large files and and do not allow the direct comparison with observations. The package provide a function that extract time-series from geogrephic locations from a site-list. The function `extract_serie()` extract the time-series from a list of files and save the output in R file (in this example the output is read). 
-```{r model}
-## to extract time-series from WRF-Chem model
-## wrf_files <- dir(pattern = "wrfout_d01")
-## extract_serie(filelist = wrf_files, point = site_list, variable="T2", prefix="mmodel.d01", field="3d")
[…]
