+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.1
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-03-26T18:19Z
+last_built: 2025-03-26T19:00Z
 urls:
   reference: https://schuch666.github.io/eva3dm/reference
   article: https://schuch666.github.io/eva3dm/articles
