+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: eva3dm
 Type: Package
 Title: Evaluation of 3D Meteorological and Air Quality Models
 Version: 0.99.1
-Date: 2024-11-24
+Date: 2025-03-26
 Description: Provides tools for post-process, evaluate and visualize results from 3d Meteorological and Air Quality models against point observations (i.e. surface stations) and grid (i.e. satellite) observations.
 Authors@R: c(
   person(given = "Daniel",
