### 2025-03-31 #7797 @Schuch666
Dear all,
I did a updated figure 1 due to the low quality of previous figure
thank you
