### 2025-04-02 #7797 @jonsampedro
Hello @Schuch666, @elbeejay,

Thanks for addressing the issues raised in my previous review and for providing detailed responses to my comments. With the new information (+figure) provided, the package’s contributions and capabilities are much clearer to me. While I acknowledge the improvements in the documentation, I still have a few minor comments.

Regarding the **documentation**, I see that a vignette has been included with an example of the package’s workflow. However, I couldn’t find this tutorial on the documentation website. I recommend adding it as a dedicated tab, similar to "_Reference_" or "_Changelog_".

 I have also some comments in the **manuscript**:
-	L14: Which “_specific aspects_”?
-	L18: Unclear “_depending on the variable_”
-	L34: This sub-step is not clear for me, as you are still describing the observational data processing.
-	L40-44: Is this information/functions related to the “Quality Assurance of the observation data”?
-	L50 and L53: typo “.”
-	L55: What are “u” and “v”
-	L60-64: Too long… divide to different sentences (for _eva_ and _sat_, respectively?)
-	L73: Typo “special”
-	L73: I think this table needs some contextualization. Why are you including these additional functions? Why are them used for?
-	L76: Could this be more elaborated? You could point to the recently added vignette…

My final point is that I cannot find the "_Contribution_" file, which, according to the response, was supposedly added to the repository.

Hope these helps. No additional comments from my side!
