### 2025-04-02 bf97ca2
Update CODE_OF_CONDUCT.md
