+++ CODE_OF_CONDUCT.md
@@ -54,12 +54,13 @@ further defined and clarified by project maintainers.
 
 ## Enforcement
 
-Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting the JOSS editor-in-chief at <arfon.smith@gmail.com>. All
-complaints will be reviewed and investigated and will result in a response that
-is deemed necessary and appropriate to the circumstances. The project team is
-obligated to maintain confidentiality with regard to the reporter of an incident.
-Further details of specific enforcement policies may be posted separately.
+Instances of abusive, harassing, or otherwise unacceptable behavior may be 
+reported by contacting the project team at <underschuch@gmail.com>. All 
+complaints will be reviewed and investigated and will result in a response 
+that is deemed necessary and appropriate to the circumstances. The project 
+team is obligated to maintain confidentiality with regard to the reporter 
+of an incident. Further details of specific enforcement policies may be 
+posted separately.
 
 Project maintainers who do not follow or enforce the Code of Conduct in good
 faith may face temporary or permanent repercussions as determined by other
