+++ CONTRIBUTING.md
@@ -2,7 +2,9 @@
 
 Thank you for any and all contributions! Following these guidelines will help streamline the process of contributing and make sure that we're all on the same page. While we ask that you read this guide and follow it to the best of your abilities, we welcome contributions from all, regardless of your level of experience.
 
-By participating in this project, you agree to abide by the [code of conduct](https://github.com/schuch666/eva3dm/blob/master/CONDUCT.md).
+By participating in this project, you agree to abide by the [code of conduct](https://github.com/schuch666/eva3dm/blob/JOSS/CODE_OF_CONDUCT.md).
+
+
 
 # Types of contributions
 
