### 2025-04-02 #7797 @Schuch666
Hello @jonsampedro , @elbeejay

I did a few extra updates and some missing points from previous review round, see bellow

In relation to the website, the website is generated based on the main branch, I will merge the final JOSS brand of the repository with the main branch after the review process and it will be included (vignette need some extra configuration in order to work properlly).

In relation the manuscript:
    L14: added a note "data visualization, geoprocessing, etc"
    L18: changed to "depending on the evaluated variable"
    L34: This step is removed, is described in next step and vignette
    L40-44: No, These functions are related to processing of time-series and conversion of units, it was moved to Process observation data for evaluation
    L50 and L53: updated
    L55: u and v are eastward and northward wind component, included in the paper
    L60-64: updated to two sentences
    L73: updated
    L73: Table updated with clear objectives
    L76: updated including a reference to the documentation and vignette

In relation to the Contribution, A session the read-me including links for the Contribution page and code of conduct page was added to the JOSS branch (it will be reflected to the website after merge the changes on JOSS branch).

Thank you, let me know if there are additional points
