### 2025-04-02 #7797 @elbeejay
Thanks for making these changes @Schuch666 and big thanks to @jonsampedro for your quick responses!

@zhenchun - have you had a chance to take a fuller look at this submission? If not, do you know when you may have some time? Thanks!
