### 2025-04-04 #7797 @elbeejay
Thanks @jonsampedro! Do you mind finishing your [reviewer checklist ](https://github.com/openjournals/joss-reviews/issues/7797#issuecomment-2670751233) now that all of your concerns have been satisfied?
