### 2025-04-08 385b24c
skip_interp argument added to sat()
