+++ R/sat.R
@@ -14,6 +14,7 @@
 #' @param method passed to terra::resample
 #' @param eval_function evaluation function (default is stat)
 #' @param mask optional SpatVector to mask the results
+#' @param skip_interp skip the interpolation step
 #' @param verbose set TRUE to display additional information
 #' @param ... other arguments passed to stat
 #'
@@ -40,7 +41,7 @@
 sat <- function(mo,ob,rname, table = NULL,
                 n = 6, min = NA, max = NA,
                 method = 'bilinear', eval_function = stat,
-                mask, verbose = TRUE, ...){
+                mask, skip_interp = FALSE, verbose = TRUE, ...){
 
   if(missing(mo))
     stop('model input is missing!') # nocov
@@ -55,7 +56,9 @@ sat <- function(mo,ob,rname, table = NULL,
     ob <- rast(ob) # nocov
   }
 
-  cut_boundary <- function(x, n,value = NA){
+  cut_boundary <- function(x, n,value = NA, verbose = FALSE){
+
+    if(verbose) cat(paste0('removing ',n,' points for the model (y) boundaryes ...\n'))
 
     if(n < 1) return(x) # nocov
 
@@ -83,10 +86,15 @@ sat <- function(mo,ob,rname, table = NULL,
     }
   }
 
-  if(verbose) cat(paste0('removing ',n,' points for the model (y) boundaryes ...\n'))
-  model <- cut_boundary(mo, n = n)
-  if(verbose) cat('interpolating obs. (x) to model grid (y)...\n')
-  obser <- interp(x = ob, y = mo, method = method, mask = mask, verbose = verbose)
+  model <- cut_boundary(mo, n = n, verbose = verbose)
+
+  if(!skip_interp){
+    if(verbose) cat('interpolating obs. (x) to model grid (y) ...\n')
+    obser <- interp(x = ob, y = mo, method = method, mask = mask, verbose = verbose)
+  }else{
+    if(verbose) cat('skiping interpolation ...\n')
+    obser <- ob
+  }
 
   if(!is.na(min)){
     if(verbose) cat('seting min value to',min,'\n')
