### 2025-04-11 #7797 @elbeejay
Thanks @zhenchun!

@Schuch666 I will generate a new checklist with some items for you and me to complete. I will be primarily reading through the paper and checking the bibliography and metadata; please let me know when you have done things like generated the archive and fixed the version -- you will not be able to edit this checklist yourself.
