### 2025-04-11 #7797 @elbeejay
Checking DOIs, I was unable to find the title, source material, or a reference to the PacWest publication. @Schuch666 did you access this online? If so, where/how? is it a publication or website or guide?

For the WRF description it looks like there is a DOI, see this [webpage](https://opensky.ucar.edu/islandora/object/technotes%3A576).
