### 2025-04-11 #7797 @elbeejay
In the paper itself, please revise:

- Line 13: switch "or" for "and"
- Line 25: suggest switching "on the next sessions" for "in the next sections"
- Figure 1 caption, please remove "Figure 1 - " from the caption text
- Line 27: suggest removing the word "of" following "Download"
- Line 34: suggest adding "the" before "NASA giovanni", and suggest capitalizing "Giovanni"
- Line 45: suggest pluralizing "variable"
- Line 68: suggest revising the sentence to "Table 1 describes key functions" or similar (as Table 1 does not exclusively describe visualization functions)
- Line 69: replace "on" with "in" or "from"
- Figure 2 caption text, remove the "Figure 2 - "
- Line 70: Typo, fix spelling of "Below"; I suggest removing "presented" and adding "the" before "evaluation" so the sentence reads: "Below is a simple example of the evaluation of temperature..."
- Line 74: consider adding "model," after the word "Chemistry"
