+++ paper.md
@@ -24,23 +24,23 @@ Eva3dm is a package designed to support the evaluation of 3-dimensional physical
 
 # Statement of need
 
-Evaluation is a crucial step in any model application, as it ensures that the model results accurately represent the variables of interest. Without a good evaluation process, the reliability and applicability of model outputs remain uncertain. There are currently other tools available in R [@David:2012], Python [@Ladwig:2017] or other languages [@NCAR:2019;@Appel:2011]. However, these tools often focus on specific aspects (data visualization, geoprocessing, etc) and lack a fully integrated framework. This package fills that gap by streamlining the entire evaluation process—from preprocessing observations and model outputs to statistical analysis and visualization—offering a comprehensive and user-friendly solution for air quality model assessment.
+Evaluation is a crucial step in any model application, as it ensures that the model results accurately represent the variables of interest. Without a good evaluation process, the reliability and applicability of model outputs remain uncertain. There are currently other tools available in R [@David:2012], Python [@Ladwig:2017] and other languages [@NCAR:2019;@Appel:2011]. However, these tools often focus on specific aspects (data visualization, geoprocessing, etc) and lack a fully integrated framework. This package fills that gap by streamlining the entire evaluation process—from preprocessing observations and model outputs to statistical analysis and visualization—offering a comprehensive and user-friendly solution for air quality model assessment.
 
 # Description
 
-The literature presents various evaluation criteria depending on the evaluated variable [@Emery:2001;@Ramboll:2018;@Monk:2019;@Zhang:2019;@Emery:2017;@Zhai:2024], which can be used to compare models and assess their performance. These criteria vary based on the simulation goal, observation variability, and measurement errors.
+The literature presents various evaluation criteria depending on the evaluated variable [@Emery:2001;@Monk:2019;@Zhang:2019;@Emery:2017;@Zhai:2024], which can be used to compare models and assess their performance. These criteria vary based on the simulation goal, observation variability, and measurement errors.
 
-A brief description of the steps to perform a model evaluation and the functions to support these steps are described on the next sessions and Figure 1 shows a diagram of the workflow.
+A brief description of the steps to perform a model evaluation and the functions to support these steps are described in the next sections and Figure 1 shows a diagram of the workflow.
 
-![Figure 1 - Diagram of the evaluation processs.\label{fig:diagramm}](joss_2.png){ width=100% }
+![Diagram of the evaluation processs.\label{fig:diagramm}](joss_2.png){ width=100% }
 
 ### 1. Pre-processing of observations
 
-- Download of the observations, some examples include: 
+- Download the observations, some examples include: 
   - METAR (METeorological Aerodrome Report) can be downloaded using the R-package [riem](https://docs.ropensci.org/riem/) or the [Iowa State University](https://mesonet.agron.iastate.edu/request/download.phtml) website
   - AERONET (Aerosol Robotic Network) can be downloaded at [AErosol RObotic NETwork](https://aeronet.gsfc.nasa.gov/new_web/data.html) website
   - Air Quality data for Brazil can be downloaded using the R-package [qualR](https://github.com/ropensci/qualR), or [QUALAR](https://qualar.cetesb.sp.gov.br/qualar) and [MonitorAir](https://www.data.rio/datasets/dados-hor%C3%A1rios-do-monitoramento-da-qualidade-do-ar-monitorar/explore) websites
-  - Satellite products are available at [NASA giovanni](https://giovanni.gsfc.nasa.gov/giovanni/) website
+  - Satellite products are available at the [NASA Giovanni](https://giovanni.gsfc.nasa.gov/giovanni/) website
 
[…]
