+++ paper.bib
@@ -1,8 +1,8 @@
-@article{Emery:2001,
-  	Author = {{Emery}, C. et. al},
-  	Journal = {Prepared for the Texas natural resource conservation commission, by ENVIRON International Corporation},
-  	Title = {{Enhanced Meteorological Modeling and Performance Evaluation for Two Texas Ozone Episodes}},
-  	Year = 2001
+@inproceedings{Emery:2001,
+  author={Christopher A. Emery and Edward Tai},
+  title={Enhanced Meteorological Modeling and Performance Evaluation for Two Texas Ozone Episodes},
+  year={2001},
+  url={https://api.semanticscholar.org/CorpusID:127579774}
 }
 
 @article{Monk:2019,
@@ -34,19 +34,14 @@
   	Journal = {Atmospheric Environment},
   	Title = {{Recommendations on benchmarks for photochemical air quality model applications in China—NO2, SO2, CO and PM10}},
   	Year = 2024,
-  	DOI = {10.1080/10962247.2016.1265027}
-}
-
-@misc{Ramboll:2018,
-  author = {{Ramboll}},
-  title = {PacWest Newport Meteorological Performance Evaluation},
-  year = {2018},
+  	DOI = {10.1016/j.atmosenv.2023.120290}
 }
 
 @misc{Skamarock:2019,
   	Author = {{Skamarock}, W. et. al},
-  	Title = {{A description of the advanced research WRF version 4}},
+  	Title = {{A Description of the Advanced Research WRF version 4.1}},
   	Year = 2024,
+  	DOI = {https://doi.org/10.5065/1dfh-6p97},
 }
 
 @article{Grell:2005,
