### 2025-04-11 #7797 @Schuch666
Hello @elbeejay thank you for the revision

The PacWest is a report, I have found online (I still have the pdf), but now, it seems that the file is no longer available online. I'm removing the reference from the paper.

I did the update of WRF DOI and include some missing DOI

All the points are fixed in the paper

Let me know if you have any additional question
