+++ paper.bib
@@ -41,7 +41,7 @@
   	Author = {{Skamarock}, W. et. al},
   	Title = {{A Description of the Advanced Research WRF version 4.1}},
   	Year = 2024,
-  	DOI = {https://doi.org/10.5065/1dfh-6p97},
+  	DOI = {10.5065/1dfh-6p97},
 }
 
 @article{Grell:2005,
