### 2025-04-14 #7797 @Schuch666
Hello @elbeejay

I just fix the Invalid DOI and did a new Zenode archive of this version

Schuch, D. (2025). eva3dm: A R-package for model evaluation of 3D weather and air quality models (1.00). Zenodo. https://doi.org/10.5281/zenodo.15213216
