### 2025-04-14 ed93a16
Update sat.R cod-cov
