### 2025-04-15 #7797 @elbeejay
@Schuch666 we are tracking version 1.00 above, but in your repository the CRAN version in the badge is noted as 0.99.1. Can you please clarify for me which version of the software should become the version of record we tie to the JOSS publication?
