### 2025-04-15 #7797 @Schuch666
Hello @elbeejay the version to be tie to the JOSS publication is 1.00 (JOSS brand).

This JOSS brand is ahead of the CRAN version

There is also the main brand, this brand is similar to the version available in CRAN (with a few small changes) and I'm planing to merge the main brand with the JOSS brand and send to CRAN in the next weeks
