### 2025-04-17 #7797 @kthyng
@Schuch666 Usually people do a release or tag of their software to associate with the JOSS submission. Do you want to do this?

Also please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example see "texas".
