+++ paper.bib
@@ -1,6 +1,6 @@
 @inproceedings{Emery:2001,
   author={Christopher A. Emery and Edward Tai},
-  title={Enhanced Meteorological Modeling and Performance Evaluation for Two Texas Ozone Episodes},
+  title={{Enhanced Meteorological Modeling and Performance Evaluation for Two Texas Ozone Episodes}},
   year={2001},
   url={https://api.semanticscholar.org/CorpusID:127579774}
 }
@@ -8,7 +8,7 @@
 @article{Monk:2019,
   	Author = {{Monk}, K. et. al},
   	Journal = {Atmosphere},
-  	Title = {{Enhanced Meteorological Modeling and Performance Evaluation for Two Texas Ozone Episodes}},
+  	Title = {{Evaluation of Regional Air Quality Models over Sydney and Australia: Part 1—Meteorological Model Comparison}},
   	Year = 2019,
   	DOI = {10.3390/atmos10070374}
 }
@@ -67,7 +67,7 @@
 }
 
 @Article{David:2012,
-    title = {openair - An {R} package for air quality data analysis},
+    title = {{openair - An R package for air quality data analysis}},
     author = {{David} C. Carslaw and Karl Ropkins},
     journal = {Environmental Modelling & Software},
     volume = {27--28},
@@ -86,14 +86,14 @@
 }
 
 @misc{NCAR:2019,
-  	Author = {{NCAR}, and UCAR and CISL and TDD},
+  	Author = {{NCAR, and UCAR and CISL and TDD}},
   	Title = {{The NCAR Command Language (Version 6.6.2)}},
   	Year = 2017,
   	doi = {10.5065/D6WD3XH5},
 }
 
 @article{Appel:2011,
-  	Author = {{Appel}, K. W., Gilliam, R. C., Davis, N., Zubrow, A., & Howard, S. C.},
+  	Author = {{Appel, K. W., Gilliam, R. C., Davis, N., Zubrow, A., & Howard, S. C.}},
   	Journal = {Environmental Modelling & Software},
   	Title = {{Overview of the Atmospheric Model Evaluation Tool (AMET) v1.1 for evaluating meteorological and air quality models}},
   	Year = 2011,
