+++ paper.bib
@@ -8,7 +8,7 @@
 @article{Monk:2019,
   	Author = {{Monk}, K. et. al},
   	Journal = {Atmosphere},
-  	Title = {{Evaluation of Regional Air Quality Models over Sydney and Australia: Part 1—Meteorological Model Comparison}},
+  	Title = {{Evaluation of Regional Air Quality Models over Sydney and Australia: Part 1-Meteorological Model Comparison}},
   	Year = 2019,
   	DOI = {10.3390/atmos10070374}
 }
@@ -32,7 +32,7 @@
 @article{Zhai:2024,
   	Author = {{Zhai}, H. et. al},
   	Journal = {Atmospheric Environment},
-  	Title = {{Recommendations on benchmarks for photochemical air quality model applications in China—NO2, SO2, CO and PM10}},
+  	Title = {{Recommendations on benchmarks for photochemical air quality model applications in China-NO2, SO2, CO and PM10}},
   	Year = 2024,
   	DOI = {10.1016/j.atmosenv.2023.120290}
 }
