### 2025-04-17 #7797 @Schuch666
Hi @kthyng thank you!

Ok, I did a new release for the package: JOSS-1.00, soon I will merge with other changes included in the main band and make a new release (for next CRAN submission)

 I fix some missing "{}" and the title of one of the papers
