### 2025-04-17 #7797 @kthyng
Ok sounds good. We're ready to wrap up then.
